cdk/lambdas/requirements.txt
psycopg[binary]==3.2.4
PyJWT[crypto]==2.10.1
cryptography==44.0.0
